backend/requirements.txt
boto3
django 
djangorestframework 
djangorestframework-simplejwt
django-cors-headers
faiss-cpu
langchain
langchain_aws
langchain_community
langgraph
matplotlib
